useMemo for plugins search
